[Feat]: HistoryView 작성 #79
