+++ ForJaeRin/Presentation/Views/ProjectView/ProjectHistoryView/MissedKeywordListView.swift
@@ -1,0 +1,28 @@
+//
+//  MissedKeywordListView.swift
+//  ForJaeRin
+//
+//  Created by Yun Dongbeom on 2023/07/27.
+//
+
+import SwiftUI
+
+struct MissedKeywordListView: View {
+    
+    var body: some View {
+        HStack(spacing: 28) {
+//            keywordSuccessView()
+        }
+        .border(.red, width: 2)
+    }
+}
+
+extension MissedKeywordListView {
+
+}
+
+//struct MissedKeywordListView_Previews: PreviewProvider {
+//    static var previews: some View {
+//        MissedKeywordListView()
+//    }
+//}
