+++ ForJaeRin/Presentation/Views/ProjectView/ProjectHistoryView/ProjectSummaryView/ProjectSummaryView.swift
@@ -61,7 +61,7 @@ struct ProjectSummaryView: View {
                                     }
                                     
                                     ZStack{
-                                        PieChart(slices : data)
+                                        PieChartView(slices : data)
                                             .foregroundColor(Color.primary500)
                                             .frame(width: 200, height: 200)
                                         
@@ -212,385 +212,6 @@ struct ProjectSummaryView: View {
 }
 
 
-
-
-// MARK: - 그룹 한 눈에 보기
-struct GroupView: View {
-    
-    let group: String
-    //[data 가져오기] 그룹 내 키워드들 텍스트 가져와야 함
-    let groupKeyword1: [String] = ["차이", "디테일", "디자이너", "개발자", "소통", "잘", "전달"]
-    let groupKeyword2: [String] = ["소통", "방법"]
-    let groupKeyword3: [String] = ["HIG", "타이포그래피짱짱맨안녕하세요반가워요짱짱", "가독성", "자동", "사이즈", "일관된", "개발자"]
-    let groupKeyword4: [String] = ["스타일", "가이드", "폰트", "이름이어려워요", "컬러", "아이콘"]
-    let groupKeyword5: [String] = ["프로토타입", "코멘트", "귀찮", "쉽게"]
-    let groupKeyword6: [String] = ["스타일", "케이스", "지정", "편리", "효율적", "공유"]
-    let groupKeyword7: [String] = ["결론적", "Xcode", "디테일"]
-    
-    //[data 가져오기] 그룹 내 키워드들을 사용자가 말했는지 여부 (true : 키워드 말함, false : 말하지 못함)
-    let groupKeyword1Feedback: [Bool] = [true, true, true, true, false, false, true]
-    let groupKeyword2Feedback: [Bool] = [false, true]
-    let groupKeyword3Feedback: [Bool] = [true, false, true, true, false, true, true]
-    let groupKeyword4Feedback: [Bool] = [true, true, true, true, false, true]
-    let groupKeyword5Feedback: [Bool] = [false, true, true, true]
-    let groupKeyword6Feedback: [Bool] = [true, true, true, false, true, true]
-    let groupKeyword7Feedback: [Bool] = [true, false, true]
-    
-    func color(for group: String) -> Color {
-        switch group {
-        case "Group 1":
-            return .groupPurple
-        case "Group 2":
-            return .groupYellow
-        case "Group 3":
-            return .groupGreen
-        case "Group 4":
-            return .groupPink
-        case "Group 5":
-            return .groupOrange
-        case "Group 6":
-            return .groupBlue
-        case "Group 7":
-            return .groupGray
-        default:
-            return .systemGray300
-        }
-    }
-    
-    func colortext(for group: String) -> Color {
-        switch group {
-        case "Group 1":
-            return .groupPurpleText
-        case "Group 2":
-            return .groupYellowText
-        case "Group 3":
-            return .groupGreenText
-        case "Group 4":
-            return .groupPinkText
-        case "Group 5":
-            return .groupOrangeText
-        case "Group 6":
-            return .groupBlueText
-        case "Group 7":
-            return .groupGrayText
-        default:
-            return .systemGray500
-        }
-    }
-    
-    func grouptext(for group: String) -> [String] {
-        switch group {
-        case "Group 1":
-            return groupKeyword1
-        case "Group 2":
-            return groupKeyword2
-        case "Group 3":
-            return groupKeyword3
-        case "Group 4":
-            return groupKeyword4
-        case "Group 5":
-            return groupKeyword5
-        case "Group 6":
-            return groupKeyword6
-        case "Group 7":
-            return groupKeyword7
-        default:
-            return []
-        }
-    }
-    
-    func groupKeywordCount(for group: String) -> Int {
-        switch group {
-        case "Group 1":
-            return groupKeyword1.count
-        case "Group 2":
-            return groupKeyword2.count
-        case "Group 3":
-            return groupKeyword3.count
-        case "Group 4":
-            return groupKeyword4.count
-        case "Group 5":
-            return groupKeyword5.count
-        case "Group 6":
-            return groupKeyword6.count
-        case "Group 7":
-            return groupKeyword7.count
-        default:
-            return 0
-        }
-    }
-    
-    func groupKeywordFeedback (for group: String) -> [Bool] {
-        switch group {
-        case "Group 1":
-            return groupKeyword1Feedback
-        case "Group 2":
-            return groupKeyword2Feedback
-        case "Group 3":
-            return groupKeyword3Feedback
-        case "Group 4":
-            return groupKeyword4Feedback
-        case "Group 5":
-            return groupKeyword5Feedback
-        case "Group 6":
-            return groupKeyword6Feedback
-        case "Group 7":
-            return groupKeyword7Feedback
-        default:
-            return []
-        }
-    }
-    
-    // 키워드 갯수에 따라 사각형의 높이를 계산하는 함수
-    //if 페이지가 2장 이상일 경우에 구분선이 추가되는데, 위에서 계산한 minHeight에 32*(페이지 갯수-1)만큼 더해주면 될 듯. 그리고 구분선 추가
-    func calculateRectangleHeight(keywordCount: Int) -> CGFloat {
-        return CGFloat(20 + 8 + 40 * groupKeywordCount(for: group) + 8 * (groupKeywordCount(for: group)-1) + 8)
-    }
-    
-    var body: some View {
-        ZStack (alignment: .topLeading) {
-            Rectangle()
-                .cornerRadius(12)
-                .frame(
-                    minWidth: 160,
-                    maxWidth: .infinity,
-                    minHeight: calculateRectangleHeight(keywordCount: groupKeywordCount(for: group)),
-                    maxHeight: calculateRectangleHeight(keywordCount: groupKeywordCount(for: group)))
-                .foregroundColor(Color.systemWhite)
-            VStack {
-                ZStack(alignment: .bottom) {
-                    Rectangle()
-                        .cornerRadius(12)
-                        .frame(minWidth: 160, maxWidth: .infinity, minHeight: 20, maxHeight: 20)
-                        .foregroundColor(color(for: group))
-                    
-                    Rectangle()
-                        .frame(minWidth: 160, maxWidth: .infinity, minHeight: 10, maxHeight: 10)
-                        .foregroundColor(color(for: group))
-                }
-                
-                //키워드 감싸는 사각형 (true면 기본 호출 & false면 피드백 호출)
-                VStack (spacing: 8){
-                    ForEach(Array(zip(grouptext(for: group), groupKeywordFeedback(for: group))), id: \.0) { keyword, feedback in
-                        ZStack {
-                            RoundedRectangle(cornerRadius: 8)
-                                .stroke(feedback ? Color.systemGray100 : colortext(for: group), lineWidth: 1)
-                                .frame(minWidth: 20, maxWidth: .infinity, minHeight: 40, maxHeight: 40)
-                                .background(Color.systemWhite)
-                            VStack {
-                                Text(keyword)
-                                    .padding(.horizontal, 10)
-                                    .frame(minWidth: 40, maxWidth: .infinity, minHeight: 40, maxHeight: 40)
-                                    .foregroundColor(feedback ? Color.systemGray400 : colortext(for: group))
-                                    .font(.systemSubtitle)
-                            }
-                        }
-                        .padding(.horizontal, 6)
-                    }
-                    .fixedSize()
-                }
-                
-            }
-            .frame(minHeight: 80, maxHeight: .infinity)
-        }
-        .frame(maxHeight: .infinity)
-    }
-}
-
-// MARK: - 그룹 차트
-struct GroupChartView: View {
-    var groupName: String
-    var realGroup: Int
-    
-    var maxTime: Int
-    var hopeTime: Int
-    var realTime: Int
-    
-    var body: some View {
-        Group {
-            VStack (spacing: 0){
-                HStack (spacing: 0){
-                    
-                    if realGroup > 0
-                    {
-                        Image(systemName: "plus")
-                            .frame(width: 16, height: 16)
-                            .padding(.trailing, 2)
-                            .foregroundColor(Color.systemPrimary)
-                        
-                        //임의로 적어둔"01:00" 부분에 RealGroup 변수 값이 들어가야 할 듯
-                        Text("01:00")
-                            .systemFont(.caption1)
-                            .foregroundColor(Color.systemPrimary)
-                    } else if realGroup < 0 {
-                        Image(systemName: "minus")
-                            .frame(width: 16, height: 16)
-                            .padding(.trailing, 2)
-                            .foregroundColor(Color.systemGray200)
-                        
-                        //임의로 적어둔 "00:20" 부분에 RealGroup 변수 값이 들어가야 할 듯
-                        Text("00:20")
-                            .systemFont(.caption1)
-                            .foregroundColor(Color.systemGray200)
-                    } else {
-                    //임의로 적어둔 "00:00" 부분에 RealGroup 변수 값이 들어가야 할 듯
-                    Text("00:00")
-                        .systemFont(.caption1)
-                        .foregroundColor(Color.systemGray200)
-                    }
-                }
-                
-                if realGroup > 0
-                {
-                    ZStack{
-                        Rectangle()
-                            .frame(width: 78, height: 219)
-                            .foregroundColor(Color.primary100)
-                            .cornerRadius(6)
-                            .opacity(100)
-                            .padding(.top, 8)
-                            .padding(.bottom, 5)
-                        
-                        StyleDetailsChart(maxTime: maxTime, hopeTime: hopeTime, realTime: realTime)
-                    }
-                } else {
-                    ZStack{
-                        Rectangle()
-                            .frame(width: 78, height: 219)
-                            .foregroundColor(Color.primary100)
-                            .cornerRadius(6)
-                            .opacity(0)
-                            .padding(.top, 8)
-                            .padding(.bottom, 5)
-                        
-                        StyleDetailsChart(maxTime: maxTime, hopeTime: hopeTime, realTime: realTime)
-                    }
-                }
-                
-                if realGroup > 0
-                {
-                    Text(groupName)
-                        .systemFont(.caption2)
-                        .foregroundColor(Color.systemGray500)
-                } else {
-                    Text(groupName)
-                        .systemFont(.caption2)
-                        .foregroundColor(Color.systemGray300)
-                }
-            }
-        }
-    }
-}
-
-
-// MARK: - 막대 그래프
-
-struct StyleDetailsChart: View {
-    var maxTime: Int
-    var hopeTime: Int
-    var realTime: Int
-
-    var body: some View {
-        HStack {
-            VStack {
-                
-                CustomBarChart1(data: [
-                    //[data 가져오기]
-                    //사용자가 설정한 그룹 발표 시간 중 가장 최대값 : maxTime
-                    //y에 사용자가 설정한 그룹 발표 시간 : hopeTime
-                    //y에 들어갈 수식(goalGraph) : 200*(hopeTime/(maxTime+(maxTime*1/4)))
-                    (x: "Cachapa", y: Int(Double(200) * (Double(hopeTime) / (Double(maxTime) + (Double(maxTime) * 0.15)))))
-                ])
-            }
-            .frame(width: 18, height: 200, alignment: .bottom)
-            
-            VStack {
-                CustomBarChart2(data: [
-                    //[data 가져오기]
-                    //y에 사용자가 실제로 말한 그룹 발표 시간 : realTime
-                    //y에 들어갈 수식(realGraph) : hopeTime-realTime
-                    (x: "Cachapa", y: Int(Double(200) * (Double(realTime) / (Double(maxTime) + (Double(maxTime) * 0.15)))))
-                ])
-            }
-            .frame(width: 18, height: 200, alignment: .bottom)
-        }
-    }
-}
-
-struct CustomBarChart1: View {
-    let data: [(x: String, y: Int)]
-    
-    var body: some View {
-        ZStack (alignment: .bottomLeading) {
-            Rectangle()
-                .frame(width: 18, height: CGFloat(data[0].y))
-                //오퍼시티 없는 색 버전으로 수정 필요
-                .foregroundColor(Color.systemGray200)
-                .cornerRadius(6)
-            Rectangle()
-                .frame(width: 18, height: 6)
-                //오퍼시티 없는 색 버전으로 수정 필요
-                .foregroundColor(Color.systemGray200)
-            ForEach(data, id: \.x) { datum in
-            }
-        }
-    }
-}
-
-struct CustomBarChart2: View {
-    let data: [(x: String, y: Int)]
-    
-    var body: some View {
-        ZStack (alignment: .bottomLeading) {
-            Rectangle()
-                .frame(width: 18, height: CGFloat(data[0].y))
-                .foregroundColor(Color.systemPrimary)
-                .cornerRadius(6)
-            Rectangle()
-                .frame(width: 18, height: 6)
-                .foregroundColor(Color.systemPrimary)
-            ForEach(data, id: \.x) { datum in
-            }
-        }
-    }
-}
-
-// MARK: - 원 그래프
-
-struct PieChart: View {
-    @State var slices: [(Double, Color)]
-    
-    var body: some View {
-        Canvas { context, size in
-            //도넛 모양
-            let donut = Path { path in
-                path.addEllipse(in: CGRect(origin: .zero, size: size))
-                path.addEllipse(in: CGRect(x: size.width * 0.1, y: size.height * 0.1, width: size.width * 0.8, height: size.height * 0.8))
-            }
-            context.clip(to: donut, style: .init(eoFill: true))
-            
-            let total = slices.reduce(0) { $0 + $1.0 }
-            context.translateBy(x: size.width * 0.5, y: size.height * 0.5)
-            var pieContext = context
-            pieContext.rotate(by: .degrees(-90))
-            let radius = min(size.width, size.height)
-
-            var startAngle = Angle.zero
-            for (value, color) in slices {
-                let angle = Angle(degrees: 360 * (value / total))
-                let endAngle = startAngle + angle
-                let path = Path { path in
-                    path.move(to: .zero)
-                    path.addArc(center: .zero, radius: radius, startAngle: startAngle + Angle(degrees: 0), endAngle: endAngle, clockwise: false)
-                    path.closeSubpath()
-                }
-                pieContext.fill(path, with: .color(color))
-                startAngle = endAngle
-            }
-        }
-        .aspectRatio(1, contentMode: .fit)
-    }
-}
 
 struct ProjectSummaryView_Previews: PreviewProvider {
     static var previews: some View {
